move books to its own file, move IBook interface to its own file
